+++ tf-modules/cumulus_ecs_service/main.tf
@@ -58,13 +58,13 @@ resource "aws_ecs_task_definition" "default" {
   resource "aws_ecs_task_definition" "fargate" {
   count = var.use_fargate ? 1 : 0
 
-  family       = "${local.full_name}-fg"
-  network_mode = "awsvpc"
+  family                   = "${local.full_name}-fg"
+  network_mode             = "awsvpc"
   requires_compatibilities = local.compatibilites
-  execution_role_arn = var.execution_role_arn
-  task_role_arn = var.task_role_arn
-  cpu                = var.cpu
-  memory             = var.memory_reservation
+  execution_role_arn       = var.execution_role_arn
+  task_role_arn            = var.task_role_arn
+  cpu                      = var.cpu
+  memory                   = var.memory_reservation
 
   container_definitions = jsonencode([
     {
@@ -134,11 +134,6 @@ resource "aws_ecs_service" "fargate" {
   }
   deployment_maximum_percent         = 100
   deployment_minimum_healthy_percent = 0
-  # TODO Re-enable tags once this warning is addressed:
-  #   The new ARN and resource ID format must be enabled to add tags to the
-  #   service. Opt in to the new format and try again.
-  #
-  # tags                               = var.tags
 }
 
 resource "aws_cloudwatch_metric_alarm" "custom" {
@@ -158,3 +153,135 @@ resource "aws_cloudwatch_metric_alarm" "custom" {
   }
   tags = var.tags
 }
+
+resource "aws_appautoscaling_target" "fargate_ecs_service_target" {
+  count = var.use_fargate ? 1 : 0
+  max_capacity       = var.fargate_max_capacity
+  min_capacity       = var.fargate_min_capacity
+  resource_id        = "service/${local.cluster_name}/${aws_ecs_service.fargate[0].name}"
+  scalable_dimension = "ecs:service:DesiredCount"
+  service_namespace  = "ecs"
+}
+
+resource "aws_appautoscaling_policy" "ecs_policy_fargate_up" {
+  count = var.use_fargate ? 1 : 0
+  name               =  "${local.full_name}-fg-upscale"
+  policy_type        = "StepScaling"
+  resource_id        = aws_appautoscaling_target.fargate_ecs_service_target[0].resource_id
+  scalable_dimension = aws_appautoscaling_target.fargate_ecs_service_target[0].scalable_dimension
+  service_namespace  = aws_appautoscaling_target.fargate_ecs_service_target[0].service_namespace
+
+  step_scaling_policy_configuration {
+    adjustment_type = "PercentChangeInCapacity"
+    cooldown = var.fargate_scaling_cooldown
+    metric_aggregation_type = "Average"
+    min_adjustment_magnitude = 1
+    step_adjustment {
+      metric_interval_lower_bound = 0
+      metric_interval_upper_bound = 30000
+      scaling_adjustment          = var.fargate_upscale_adjustment
+    }
+    step_adjustment {
+      metric_interval_lower_bound = 30000
+      metric_interval_upper_bound = 120000
+      scaling_adjustment          = var.fargate_upscale_adjustment * 2
+    }
+    step_adjustment {
+      metric_interval_lower_bound = 120000
+      scaling_adjustment          = var.fargate_upscale_adjustment * 5
+    }
+  }
+}
+
+resource "aws_appautoscaling_policy" "ecs_policy_fargate_down" {
+  count = var.use_fargate ? 1 : 0
+  name               = "${local.full_name}-fg-downscale"
+  policy_type        = "StepScaling"
+  resource_id        = aws_appautoscaling_target.fargate_ecs_service_target[0].resource_id
+  scalable_dimension = aws_appautoscaling_target.fargate_ecs_service_target[0].scalable_dimension
+  service_namespace  = aws_appautoscaling_target.fargate_ecs_service_target[0].service_namespace
+
+  step_scaling_policy_configuration {
+    adjustment_type = "PercentChangeInCapacity"
+    cooldown = var.fargate_scaling_cooldown
+    metric_aggregation_type = "Average"
+    min_adjustment_magnitude = 1
+    step_adjustment {
+      scaling_adjustment          = var.fargate_downscale_adjustment
+      metric_interval_upper_bound = 0
+    }
+  }
+}
+
+resource "aws_cloudwatch_metric_alarm" "scale_up" {
+  count = var.use_fargate ? 1 : 0
+  alarm_name                = "${local.full_name}-queueThreshold-scale-up"
+  comparison_operator       = "GreaterThanOrEqualToThreshold"
+  evaluation_periods        = 1
+  metric_name               = "ActivityScheduleTime"
+  namespace                 = "AWS/States"
+  period                    = var.fargate_scaling_adjustment_period
+  statistic                 = "Average"
+  threshold                 = var.fargate_scheduled_task_threshold * 1000
+  alarm_description         = "This metric monitors a fargate task activity for activity schedule time increases"
+  alarm_actions             = [aws_appautoscaling_policy.ecs_policy_fargate_up[0].arn]
+  dimensions                = {
+    ActivityArn = var.environment.ACTIVITY_ARN
+  }
+}
+
+
+resource "aws_cloudwatch_metric_alarm" "scale_down" {
+  count = var.use_fargate ? 1 : 0
+  alarm_name                = "${local.full_name}-queueThreshold-scale-down"
+  comparison_operator       = "LessThanThreshold"
+  evaluation_periods        = 1
+  metric_name               = "ActivityScheduleTime"
+  namespace                 = "AWS/States"
+  period                    = var.fargate_scaling_adjustment_period
+  statistic                 = "Average"
+  threshold                 = var.fargate_scheduled_task_threshold * 1000
+  alarm_description         = "This metric monitors a fargate task activity for a reduction in queued activities"
+  alarm_actions             = [aws_appautoscaling_policy.ecs_policy_fargate_down[0].arn]
+  dimensions                = {
+    ActivityArn = var.environment.ACTIVITY_ARN
+  }
+}
+
+resource "aws_appautoscaling_policy" "ecs_policy_fargate_off" {
+  count = var.use_fargate ? 1 : 0
+  name               = "${local.full_name}-fg-downscale"
+  policy_type        = "StepScaling"
+  resource_id        = aws_appautoscaling_target.fargate_ecs_service_target[0].resource_id
+  scalable_dimension = aws_appautoscaling_target.fargate_ecs_service_target[0].scalable_dimension
+  service_namespace  = aws_appautoscaling_target.fargate_ecs_service_target[0].service_namespace
+
+  step_scaling_policy_configuration {
+    adjustment_type = "PercentChangeInCapacity"
+    cooldown = var.fargate_scaling_cooldown
+    metric_aggregation_type = "Average"
+    min_adjustment_magnitude = 1
+    step_adjustment {
+      scaling_adjustment          = var.fargate_downscale_adjustment
+      metric_interval_upper_bound = 0
+    }
+  }
+}
+
+resource "aws_cloudwatch_metric_alarm" "service_off" {
+  count = var.use_fargate ? 1 : 0
+  alarm_name                = "${local.full_name}-queueThreshold-scale-down"
+  comparison_operator       = "LessThanThreshold"
+  evaluation_periods        = 1
+  metric_name               = "ActivitiesScheduled"
+  namespace                 = "AWS/States"
+  period                    = var.fargate_scaling_adjustment_period
+  statistic                 = "Sum"
+  threshold                 = 1
+  alarm_description         = "This alarm monitors if *any* requests have come in for the service"
+  alarm_actions             = [aws_appautoscaling_policy.ecs_policy_fargate_off[0].arn]
+  dimensions                = {
+    ActivityArn = var.environment.ACTIVITY_ARN
+  }
+}
+